+++ src/input/mod.rs
@@ -1,8 +1,22 @@
-mod mouse_tracker;
-pub(crate) use mouse_tracker::MouseTracker;
+pub mod combo;
+pub mod devices;
+pub mod relative_mouse_movement;
+pub mod shared_memory;
 
-mod keybind_receivers;
-pub(crate) use keybind_receivers::start_keybind_receivers;
+use crate::input::{
+    combo::combo_watcher, devices::Devices,
+    relative_mouse_movement::relative_mouse_movement_watcher,
+};
 
-mod devices;
-pub(crate) use devices::Devices;
+pub fn start() {
+    let devices = Devices::new();
+    let main_keyboard = devices
+        .get_main_keyboard()
+        .expect("Failed to find the main keyboard");
+    let main_mouse = devices
+        .get_main_mouse()
+        .expect("Failed to find the main mouse");
+
+    combo_watcher(main_keyboard);
+    relative_mouse_movement_watcher(main_mouse);
+}
