+++ src/lib.rs
@@ -1,11 +1,9 @@
 mod gui;
 pub use gui::run_eframe;
 
-mod input;
-
 mod mouse;
 pub use mouse::start;
 
 mod sweep;
 
-pub mod input_backend;
+pub mod input;
